Rewrite toggleable hood using RW 1.5 pawn rendering system
The hood system rewrite in v1.2.0 used dummy defs to render the hood as a separate piece of apparel, and various Harmony patches to hide the hood from the pawn inventory list. This caused various problems with general game performance and apparel restrictions.

The new system (workshopped in collaboration with Zeracronius on Discord) is re-implemented using RW 1.5's pawn render node system, with a new hood apparel graphic node dynamically added and removed, and should be much more performant.